test: migrate node status machine tests, move Node tests into own dir
