+++ packages/zwave-js/src/lib/test/node/Node.ccVersions.test.ava.ts
@@ -1,10 +1,10 @@
 import { CommandClasses } from "@zwave-js/core";
 import { MockController } from "@zwave-js/testing";
 import ava, { type TestFn } from "ava";
-import { createDefaultMockControllerBehaviors } from "../../Utils";
-import type { Driver } from "../driver/Driver";
-import { createAndStartTestingDriver } from "../driver/DriverMock";
-import { ZWaveNode } from "./Node";
+import { createDefaultMockControllerBehaviors } from "../../../Utils";
+import type { Driver } from "../../driver/Driver";
+import { createAndStartTestingDriver } from "../../driver/DriverMock";
+import { ZWaveNode } from "../../node/Node";
 
 interface TestContext {
 	driver: Driver;
