+++ packages/zwave-js/src/lib/test/node/Node.getEndpoint.test.ava.ts
@@ -6,11 +6,11 @@ import {
 import type { ThrowingMap } from "@zwave-js/shared";
 import { MockController } from "@zwave-js/testing";
 import ava, { type TestFn } from "ava";
-import { createDefaultMockControllerBehaviors } from "../../Utils";
-import type { Driver } from "../driver/Driver";
-import { createAndStartTestingDriver } from "../driver/DriverMock";
-import { DeviceClass } from "./DeviceClass";
-import { ZWaveNode } from "./Node";
+import { createDefaultMockControllerBehaviors } from "../../../Utils";
+import type { Driver } from "../../driver/Driver";
+import { createAndStartTestingDriver } from "../../driver/DriverMock";
+import { DeviceClass } from "../../node/DeviceClass";
+import { ZWaveNode } from "../../node/Node";
 
 interface TestContext {
 	driver: Driver;
